third_party/WebKit/LayoutTests/external/wpt/resource-timing/resource_timing_buffer_full_when_shrink_buffer_size-expected.txt

This is a testharness.js-based test.
PASS There are 4 scripts, and setResourceTimingBufferSize does not reduce the size.
FAIL onresourcetimingbufferfull should not be invoked during setResourceTimingBufferSize. assert_equals: onresourcetimingbufferfull should not be invoked during setResourceTimingBufferSize. expected 0 but got 2
Harness: the test ran to completion.
